api/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-cosmos==4.2.0
azure-core==1.1.1
certifi==2019.11.28
chardet==3.0.4
idna==2.8
requests==2.22.0
six==1.13.0
urllib3==1.25.7
virtualenv==16.7.9
virtualenv-clone==0.5.3
PyJWT==2.3.0
azure-identity==1.7.1
cffi==1.15.0
cryptography==36.0.1
msal==1.16.0
msal-extensions==0.3.1
portalocker==2.3.2
pycparser==2.21
azure-common==1.1.27
azure-keyvault-secrets==4.3.0
isodate==0.6.1
msrest==0.6.21
oauthlib==3.1.1
requests-oauthlib==1.3.0